dedup usage report
